Convert to pyproject
