+++ README.md
@@ -1,2 +1,11 @@
 # s3dump
-Easily list and dump open s3 buckets
+
+Easily list and dump open s3 buckets.
+
+```
+# install from source
+pip install -e .
+
+# usage
+s3d -u https://bucket-url/ -d
+```
