+++ pyproject.toml
@@ -1,0 +1,17 @@
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "s3dump"
+version = "0.1.0"
+description = "Easily list and dump open s3 buckets"
+readme = "README.md"
+requires-python = ">=3.8"
+dependencies = [
+    "requests",
+    "tqdm",
+]
+
+[project.scripts]
+s3d = "s3dump.cli:main"
